moved integration tests to test_integrations.py
