+++ tests/test_items.py
@@ -1,16 +1,4 @@
-import app, Trello_items, pytest, datetime, hypothesis
-from dotenv import find_dotenv, load_dotenv
-
-@pytest.fixture
-def client():
-    # Use our test integration config instead of the 'real' version
-    file_path = find_dotenv('.env.test')
-    load_dotenv(file_path, override=True)
-    # Create the new app.
-    test_app = app.create_app()
-    # Use the app to create a test_client that can be used in our tests.
-    with test_app.test_client() as client:
-        yield client
+import app, Trello_items, pytest, datetime
 
 #list containing a mixture of items
 TEST_LIST1 = [
